requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
multidict==6.0.4
paramiko==3.3.1
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pytz==2023.3.post1
redis==5.0.1
rest-framework-simplejwt==0.0.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
yarl==1.9.2
zope.interface==6.1
requests==2.31.0
